update Usecase orderShoe

+++ webbangiaydep/src/main/java/business/entity/CartItem.java
@@ -3,12 +3,12 @@ package business.entity;
 
 public class CartItem 
 {
-    private int productId;
-    private int variantId;
+    private Long productId;
+    private Long variantId;
     private int quantity;
     private double price;
 
-    public CartItem(int productId, int variantId, int quantity, double price) {
+    public CartItem(Long productId, Long variantId, int quantity, double price) {
         this.productId = productId;
         this.variantId = variantId;
         this.quantity = quantity;
@@ -17,8 +17,8 @@ public class CartItem
 
     public void setQuantity(int quantity) { this.quantity = quantity; }
 
-    public int getProductId() { return productId; }
-    public int getVariantId() { return variantId; }
+    public Long getProductId() { return productId; }
+    public Long getVariantId() { return variantId; }
     public int getQuantity() { return quantity; }
     public double getPrice() { return price; }
